+++ FEM-python/beam1d-python/PrePost.py
@@ -74,12 +74,12 @@ def create_model_json(DataFile):
 	# define the mesh
 	model.x = np.array(FEData['x'])
 	model.y = np.array(FEData['y'])
-	model.IEN = np.array(FEData['IEN'], np.int)
+	model.IEN = np.array(FEData['IEN'], int)
 	model.leng = model.x[model.IEN[1,:]-1] - \
                  model.x[model.IEN[0,:]-1]
 
-	model.ID = np.zeros(model.neq, np.int)
-	model.LM = np.zeros((model.neqe, model.nel), np.int)
+	model.ID = np.zeros(model.neq, int)
+	model.LM = np.zeros((model.neqe, model.nel), int)
 
 	# generate LM and ID arrays
 	setup_ID_LM()
